ddprofilingextension: degrade to HTTP where a unix socket cannot work
The socket tests created their listener under /tmp with no build constraint,
so they failed on Windows -- which this module is tested on -- before the
extension was ever exercised. Moving them behind //go:build !windows follows
what every other unix socket test in the repo does (comp/dogstatsd/listeners,
comp/api/api/apiimpl/listener) and is what makes /tmp safe to assume.

Fixing the tests exposed the more interesting question of what the extension
should do on a platform that has no socket to talk to. A collector config or a
DD_PROFILING_UNIX_SOCKET set across a mixed fleet reaches a Windows host the
same way it reaches a Linux one, and there the setting is actively harmful:
WithUDS only installs a dialer, so a path nothing serves fails once per
profiling period rather than at Start, and startForAgent has already skipped
the local forwarding server on the assumption the socket replaces it. The
result would be profiles that silently go nowhere.

So gate it. hasUnixSocketSupport() splits along the same line the rest of the
agent uses for unix socket code, and unixSocket() drops a configured socket
where it is unsupported, warns once, and lets the caller take the HTTP path it
would have taken had the setting been absent. No Windows Agent serves an APM
socket anyway: apm_config.receiver_socket has no default there and
fixupLinuxSockets, the only code that fills it in, runs on linux and aix alone.

Also nil-check e.log. NewFactory builds the extension with no agent components,
so the logger is nil in standalone mode and the new warning would have panicked
where it matters most -- the path that has no forwarding server to fall back on.

Verified with the package tests on darwin and with
GOOS=windows dda inv linter.go, which typechecks the Windows half.
