+++ comp/otelcol/ddprofilingextension/impl/unixsocket_nix.go
@@ -1,0 +1,15 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2024-present Datadog, Inc.
+
+//go:build !windows
+
+// Package ddprofilingextensionimpl defines the OpenTelemetry Profiling implementation
+package ddprofilingextensionimpl
+
+// hasUnixSocketSupport reports whether profiles can be sent to a trace agent
+// over a unix socket on this platform.
+func hasUnixSocketSupport() bool {
+	return true
+}
